# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rental Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leases:
    get:
      operationId: listLeases
      tags: [leases]
      summary: List all leases
      responses:
        "200":
          description: List of leases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lease"
    post:
      operationId: createLease
      tags: [leases]
      summary: Create a lease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
  /leases/{id}:
    get:
      operationId: getLease
      tags: [leases]
      summary: Get lease by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lease
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
    put:
      operationId: updateLease
      tags: [leases]
      summary: Update a lease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
    delete:
      operationId: deleteLease
      tags: [leases]
      summary: Delete a lease
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Payments ──────────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update/toggle payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Maintenance ──────────────────────────────────────────────────────────────
  /maintenance:
    get:
      operationId: listMaintenanceTickets
      tags: [maintenance]
      summary: List all maintenance tickets
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceTicket"
    post:
      operationId: createMaintenanceTicket
      tags: [maintenance]
      summary: Create a maintenance ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceTicketBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
  /maintenance/{id}:
    get:
      operationId: getMaintenanceTicket
      tags: [maintenance]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
    put:
      operationId: updateMaintenanceTicket
      tags: [maintenance]
      summary: Update a maintenance ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaintenanceTicketBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicket"
    delete:
      operationId: deleteMaintenanceTicket
      tags: [maintenance]
      summary: Delete a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reports ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lease:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        unitId:
          type: integer
        unitNumber:
          type: string
        propertyName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyRent:
          type: number
        deposit:
          type: number
        status:
          type: string
          enum: [active, expired, terminated, pending]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, unitId, startDate, endDate, monthlyRent, deposit, status, createdAt]
    CreateLeaseBody:
      type: object
      properties:
        tenantId:
          type: integer
        unitId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyRent:
          type: number
        deposit:
          type: number
        status:
          type: string
          enum: [active, expired, terminated, pending]
        notes:
          type: string
      required: [tenantId, unitId, startDate, endDate, monthlyRent, deposit, status]

    # Payment
    Payment:
      type: object
      properties:
        id:
          type: integer
        leaseId:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        unitNumber:
          type: string
        propertyName:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [paid, pending, overdue]
        method:
          type: string
          enum: [cash, bank_transfer, check, card, other]
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, leaseId, tenantId, amount, dueDate, status, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        leaseId:
          type: integer
        tenantId:
          type: integer
        amount:
          type: number
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [paid, pending, overdue]
        method:
          type: string
          enum: [cash, bank_transfer, check, card, other]
          nullable: true
        notes:
          type: string
          nullable: true
      required: [leaseId, tenantId, amount, dueDate, status]
    UpdatePaymentBody:
      type: object
      properties:
        status:
          type: string
          enum: [paid, pending, overdue]
        paidDate:
          type: string
          format: date
          nullable: true
        method:
          type: string
          enum: [cash, bank_transfer, check, card, other]
          nullable: true
        notes:
          type: string
          nullable: true
        amount:
          type: number
        dueDate:
          type: string
          format: date

    # Maintenance
    MaintenanceTicket:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        tenantName:
          type: string
          nullable: true
        unitId:
          type: integer
          nullable: true
        unitNumber:
          type: string
          nullable: true
        propertyName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        category:
          type: string
          enum: [plumbing, electrical, hvac, appliance, structural, other]
        assignedTo:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, priority, status, category, createdAt]
    CreateMaintenanceTicketBody:
      type: object
      properties:
        tenantId:
          type: integer
          nullable: true
        unitId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        category:
          type: string
          enum: [plumbing, electrical, hvac, appliance, structural, other]
        assignedTo:
          type: string
          nullable: true
      required: [title, description, priority, status, category]
    UpdateMaintenanceTicketBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        category:
          type: string
          enum: [plumbing, electrical, hvac, appliance, structural, other]
        assignedTo:
          type: string
          nullable: true
        tenantId:
          type: integer
          nullable: true
        unitId:
          type: integer
          nullable: true

    # Reports
...