# lib/api-zod/src/generated/types/createLeaseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rental Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeaseBodyStatus } from "./createLeaseBodyStatus";

export interface CreateLeaseBody {
  tenantId: number;
  unitId: number;
  startDate: Date;
  endDate: Date;
  monthlyRent: number;
  deposit: number;
  status: CreateLeaseBodyStatus;
  notes?: string;
}